+++ __init__.py
@@ -181,6 +181,9 @@ def register():
     from .sync_operators import GN_ImportState
     bpy.types.Scene.gnt_import_state = bpy.props.PointerProperty(type=GN_ImportState)
 
+    from .sync_ui import GN_AuditState
+    bpy.types.Scene.gnt_audit_state = bpy.props.PointerProperty(type=GN_AuditState)
+
     bpy.app.handlers.load_post.append(_on_load_post)
     bpy.app.handlers.save_post.append(_on_save_post)
     bpy.app.handlers.undo_post.append(_on_undo_post)
@@ -201,6 +204,7 @@ def unregister():
     del bpy.types.Scene.gnt_sync_prefs
     del bpy.types.Scene.gnt_commit_review
     del bpy.types.Scene.gnt_import_state
+    del bpy.types.Scene.gnt_audit_state
 
     for cls in reversed(_all_classes):
         bpy.utils.unregister_class(cls)